requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4

sentence-transformers==2.6.1
transformers==4.39.3
torch==2.2.2
numpy==1.26.4
faiss-cpu==1.13.2
scikit-learn==1.4.1.post1

pdfplumber==0.10.3
python-docx==1.1.0

requests==2.31.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
streamlit==1.33.0
typing-extensions==4.10.0